Convert remaining patterns to opengl-es 3.1

+++ patterns/blades.glsl
@@ -1,10 +1,9 @@
-#version 330 core
-
+#version 310 es
 #ifdef GL_ES
-precision mediump float;
+precision highp float;
 #endif
 
-#extension GL_OES_standard_derivatives : enable
+///// #extension GL_OES_standard_derivatives : enable
 
 uniform float time;
 uniform vec2 mouse;
