+++ patterns/fire.glsl
@@ -1,10 +1,10 @@
-#version 330 core
-
+#version 310 es
 #ifdef GL_ES
-precision mediump float;
+precision highp float;
 #endif
+
 // http://glslsandbox.com/e#30671.1
-#extension GL_OES_standard_derivatives : enable
+// #extension GL_OES_standard_derivatives : enable
 
 uniform float time;
 uniform vec2 mouse;
